pyproject.toml
[tool.poetry]
name = "DE Project2"
version = "1.0.0"
description = ""
authors = ["Mahsa Khodabandehlou"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
kafka-python = "2.0.2"
psycopg2 = "2.9.9"
python-dotenv = "1.0.1"
requests = "2.31.0"
Faker = "24.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
